[IMP] account,l10n_fi: Allow custom COA in accounting test suite
- A custom COA could be passed as parameter.
- Merge the invoice setup to the generic suite like it is in master.

closes odoo/odoo#52140

Related: odoo/enterprise#10855
