+++ addons/account/tests/account_test_savepoint.py
@@ -3,23 +3,33 @@ from odoo import fields
 from odoo.tests.common import Form, SavepointCase
 from odoo.tests import tagged
 
+from contextlib import contextmanager
+from unittest.mock import patch
+
 
 @tagged('post_install', '-at_install')
-class AccountingSavepointCase(SavepointCase):
-
-    # -------------------------------------------------------------------------
-    # DATA GENERATION
-    # -------------------------------------------------------------------------
+class AccountTestInvoicingCommon(SavepointCase):
 
     @classmethod
-    def setUpClass(cls):
-        super(AccountingSavepointCase, cls).setUpClass()
+    def copy_account(cls, account):
+        suffix_nb = 1
+        while True:
+            new_code = '%s (%s)' % (account.code, suffix_nb)
+            if account.search_count([('company_id', '=', account.company_id.id), ('code', '=', new_code)]):
+                suffix_nb += 1
+            else:
+                return account.copy(default={'code': new_code})
 
-        chart_template = cls.env.user.company_id.chart_template_id
-        if not chart_template:
+    @classmethod
+    def setUpClass(cls, chart_template_ref=None):
+        super().setUpClass()
+
+        if chart_template_ref:
+            chart_template = cls.env.ref(chart_template_ref)
+        else:
             chart_template = cls.env.ref('l10n_generic_coa.configurable_chart_template', raise_if_not_found=False)
         if not chart_template:
-            cls.skipTest("Accounting Tests skipped because the user's company has no chart of accounts.")
+            cls.skipTest(cls, "Accounting Tests skipped because the user's company has no chart of accounts.")
 
         # Create user.
         user = cls.env['res.users'].create({
@@ -34,12 +44,129 @@ class AccountingSavepointCase(SavepointCase):
         cls.env = cls.env(user=user)
         cls.cr = cls.env.cr
 
-        cls.company_data_2 = cls.setup_company_data('company_2_data')
-        cls.company_data = cls.setup_company_data('company_1_data')
+        cls.company_data_2 = cls.setup_company_data('company_2_data', chart_template)
+        cls.company_data = cls.setup_company_data('company_1_data', chart_template)
+
+        user.write({
+            'company_ids': [(6, 0, (cls.company_data['company'] + cls.company_data_2['company']).ids)],
+            'company_id': cls.company_data['company'].id,
+        })
+
         cls.currency_data = cls.setup_multi_currency_data()
 
+        # ==== Taxes ====
+        cls.tax_sale_a = cls.company_data['default_tax_sale']
+        cls.tax_sale_b = cls.company_data['default_tax_sale'].copy()
+        cls.tax_purchase_a = cls.company_data['default_tax_purchase']
+        cls.tax_purchase_b = cls.company_data['default_tax_purchase'].copy()
+        cls.tax_armageddon = cls.setup_armageddon_tax('complex_tax', cls.company_data)
+
+        # ==== Products ====
+        cls.product_a = cls.env['product.product'].create({
+            'name': 'product_a',
+            'uom_id': cls.env.ref('uom.product_uom_unit').id,
+            'lst_price': 1000.0,
+            'standard_price': 800.0,
+            'property_account_income_id': cls.company_data['default_account_revenue'].id,
+            'property_account_expense_id': cls.company_data['default_account_expense'].id,
+            'taxes_id': [(6, 0, cls.tax_sale_a.ids)],
+            'supplier_taxes_id': [(6, 0, cls.tax_purchase_a.ids)],
+        })
+        cls.product_b = cls.env['product.product'].create({
+            'name': 'product_b',
+            'uom_id': cls.env.ref('uom.product_uom_dozen').id,
+            'lst_price': 200.0,
+            'standard_price': 160.0,
+            'property_account_income_id': cls.copy_account(cls.company_data['default_account_revenue']).id,
+            'property_account_expense_id': cls.copy_account(cls.company_data['default_account_expense']).id,
+            'taxes_id': [(6, 0, (cls.tax_sale_a + cls.tax_sale_b).ids)],
+            'supplier_taxes_id': [(6, 0, (cls.tax_purchase_a + cls.tax_purchase_b).ids)],
+        })
+
+        # ==== Fiscal positions ====
+        cls.fiscal_pos_a = cls.env['account.fiscal.position'].create({
+            'name': 'fiscal_pos_a',
+            'tax_ids': [
+                (0, None, {
+                    'tax_src_id': cls.tax_sale_a.id,
+                    'tax_dest_id': cls.tax_sale_b.id,
+                }),
+                (0, None, {
+                    'tax_src_id': cls.tax_purchase_a.id,
+                    'tax_dest_id': cls.tax_purchase_b.id,
+                }),
+            ],
+            'account_ids': [
+                (0, None, {
+                    'account_src_id': cls.product_a.property_account_income_id.id,
+                    'account_dest_id': cls.product_b.property_account_income_id.id,
+                }),
+                (0, None, {
+                    'account_src_id': cls.product_a.property_account_expense_id.id,
+                    'account_dest_id': cls.product_b.property_account_expense_id.id,
+                }),
+            ],
+        })
+
+        # ==== Payment terms ====
+        cls.pay_terms_a = cls.env.ref('account.account_payment_term_immediate')
+        cls.pay_terms_b = cls.env['account.payment.term'].create({
+            'name': '30% Advance End of Following Month',
+            'note': 'Payment terms: 30% Advance End of Following Month',
+            'line_ids': [
+                (0, 0, {
+                    'value': 'percent',
+                    'value_amount': 30.0,
+                    'sequence': 400,
+                    'days': 0,
+                    'option': 'day_after_invoice_date',
+                }),
+                (0, 0, {
+                    'value': 'balance',
+                    'value_amount': 0.0,
+                    'sequence': 500,
+                    'days': 31,
+                    'option': 'day_following_month',
+                }),
+            ],
+        })
+
+        # ==== Partners ====
+        cls.partner_a = cls.env['res.partner'].create({
+            'name': 'partner_a',
+            'property_payment_term_id': cls.pay_terms_a.id,
+            'property_supplier_payment_term_id': cls.pay_terms_a.id,
+            'property_account_receivable_id': cls.company_data['default_account_receivable'].id,
+            'property_account_payable_id': cls.company_data['default_account_payable'].id,
+            'company_id': False,
+        })
+        cls.partner_b = cls.env['res.partner'].create({
+            'name': 'partner_b',
+            'property_payment_term_id': cls.pay_terms_b.id,
+            'property_supplier_payment_term_id': cls.pay_terms_b.id,
+            'property_account_position_id': cls.fiscal_pos_a.id,
+            'property_account_receivable_id': cls.company_data['default_account_receivable'].copy().id,
+            'property_account_payable_id': cls.company_data['default_account_payable'].copy().id,
+            'company_id': False,
+        })
+
+        # ==== Cash rounding ====
+        cls.cash_rounding_a = cls.env['account.cash.rounding'].create({
+            'name': 'add_invoice_line',
+            'rounding': 0.05,
+            'strategy': 'add_invoice_line',
+            'account_id': cls.copy_account(cls.company_data['default_account_expense']).id,
+            'rounding_method': 'UP',
+        })
+        cls.cash_rounding_b = cls.env['account.cash.rounding'].create({
+            'name': 'biggest_tax',
+            'rounding': 0.05,
+            'strategy': 'biggest_tax',
+            'rounding_method': 'DOWN',
+        })
+
     @classmethod
-    def setup_company_data(cls, company_name, **kwargs):
+    def setup_company_data(cls, company_name, chart_template, **kwargs):
         ''' Create a new company having the name passed as parameter.
         A chart of accounts will be installed to this company: the same as the current company one.
         The current user will get access to this company.
@@ -47,20 +174,30 @@ class AccountingSavepointCase(SavepointCase):
         :param company_name: The name of the company.
         :return: A dictionary will be returned containing all relevant accounting data for testing.
         '''
-        chart_template = cls.env.user.company_id.chart_template_id
+        def search_account(company, chart_template, field_name, domain):
+            template_code = chart_template[field_name].code
+            domain = [('company_id', '=', company.id)] + domain
+
+            account = None
+            if template_code:
+                account = cls.env['account.account'].search(domain + [('code', '=like', template_code + '%')], limit=1)
+
+            if not account:
+                account = cls.env['account.account'].search(domain, limit=1)
+            return account
+
+        currency = cls.env.user.company_id.currency_id
         company = cls.env['res.company'].create({
             'name': company_name,
-            'currency_id': cls.env.user.company_id.currency_id.id,
+            'currency_id': currency.id,
             **kwargs,
         })
         cls.env.user.company_ids |= company
-        cls.env.user.company_id = company
 
-        chart_template = cls.env['account.chart.template'].browse(chart_template.id)
-        chart_template.try_loading()
+        chart_template.try_loading(company=company)
 
         # The currency could be different after the installation of the chart template.
-        company.write({'currency_id': kwargs.get('currency_id', cls.env.user.company_id.currency_id.id)})
+        company.write({'currency_id': kwargs.get('currency_id', currency.id)})
 
         return {
             'company': company,
@@ -73,13 +210,16 @@ class AccountingSavepointCase(SavepointCase):
                     ('company_id', '=', company.id),
                     ('user_type_id', '=', cls.env.ref('account.data_account_type_expenses').id)
                 ], limit=1),
-            'default_account_receivable': cls.env['account.account'].search([
-                    ('company_id', '=', company.id),
-                    ('user_type_id.type', '=', 'receivable')
-                ], limit=1),
+            'default_account_receivable': search_account(company, chart_template, 'property_account_receivable_id', [
+                ('user_type_id.type', '=', 'receivable')
+            ]),
             'default_account_payable': cls.env['account.account'].search([
                     ('company_id', '=', company.id),
                     ('user_type_id.type', '=', 'payable')
+                ], limit=1),
+            'default_account_assets': cls.env['account.account'].search([
+                    ('company_id', '=', company.id),
+                    ('user_type_id', '=', cls.env.ref('account.data_account_type_current_assets').id)
                 ], limit=1),
             'default_account_tax_sale': company.account_sale_tax_id.mapped('invoice_repartition_line_ids.account_id'),
             'default_account_tax_purchase': company.account_purchase_tax_id.mapped('invoice_repartition_line_ids.account_id'),
@@ -108,29 +248,30 @@ class AccountingSavepointCase(SavepointCase):
         }
 
     @classmethod
-    def setup_multi_currency_data(cls):
-        gold_currency = cls.env['res.currency'].create({
+    def setup_multi_currency_data(cls, default_values={}, rate2016=3.0, rate2017=2.0):
+        foreign_currency = cls.env['res.currency'].create({
             'name': 'Gold Coin',
             'symbol': '☺',
             'rounding': 0.001,
             'position': 'after',
             'currency_unit_label': 'Gold',
             'currency_subunit_label': 'Silver',
+            **default_values,
         })
         rate1 = cls.env['res.currency.rate'].create({
             'name': '2016-01-01',
-            'rate': 3.0,
-            'currency_id': gold_currency.id,
+            'rate': rate2016,
+            'currency_id': foreign_currency.id,
             'company_id': cls.env.company.id,
         })
         rate2 = cls.env['res.currency.rate'].create({
             'name': '2017-01-01',
-            'rate': 2.0,
-            'currency_id': gold_currency.id,
+            'rate': rate2017,
+            'currency_id': foreign_currency.id,
             'company_id': cls.env.company.id,
         })
         return {
-            'currency': gold_currency,
+            'currency': foreign_currency,
             'rates': rate1 + rate2,
         }
 
@@ -213,3 +354,43 @@ class AccountingSavepointCase(SavepointCase):
                 }),
             ],
         })
+
+    @classmethod
+    def init_invoice(cls, move_type, partner=None, invoice_date=None):
+        move_form = Form(cls.env['account.move'].with_context(default_type=move_type))
+        move_form.invoice_date = invoice_date or fields.Date.from_string('2019-01-01')
+        move_form.partner_id = partner or cls.partner_a
+        with move_form.invoice_line_ids.new() as line_form:
+            line_form.product_id = cls.product_a
+        with move_form.invoice_line_ids.new() as line_form:
+            line_form.product_id = cls.product_b
+        return move_form.save()
+
+    def assertInvoiceValues(self, move, expected_lines_values, expected_move_values):
+        def sort_lines(lines):
+            return lines.sorted(lambda line: (line.exclude_from_invoice_tab, not bool(line.tax_line_id), line.name or '', line.balance))
+        self.assertRecordValues(sort_lines(move.line_ids.sorted()), expected_lines_values)
+        self.assertRecordValues(sort_lines(move.invoice_line_ids.sorted()), expected_lines_values[:len(move.invoice_line_ids)])
+        self.assertRecordValues(move, [expected_move_values])
+
+    @contextmanager
+    def mocked_today(self, forced_today):
+        ''' Helper to make easily a python "with statement" mocking the "today" date.
+        :param forced_today:    The expected "today" date as a str or Date object.
+        :return:                An object to be used like 'with self.mocked_today(<today>):'.
+        '''
+
+        if isinstance(forced_today, str):
+            forced_today = fields.Date.from_string(forced_today)
+
+        def today(*args, **kwargs):
+            return forced_today
+
+        with patch.object(fields.Date, 'today', today):
+            with patch.object(fields.Date, 'context_today', today):
+                yield
+
+
+class AccountingSavepointCase(AccountTestInvoicingCommon):
+    # Ensure the backward-compatibility before saas-13.2.
+    pass
