+++ addons/account/tests/test_account_tax.py
@@ -1,15 +1,15 @@
 # -*- coding: utf-8 -*-
-from odoo.addons.account.tests.account_test_savepoint import AccountingSavepointCase
+from odoo.addons.account.tests.account_test_savepoint import AccountTestInvoicingCommon
 from odoo.tests import tagged
 from odoo.exceptions import UserError
 
 
 @tagged('post_install', '-at_install')
-class TestAccountTax(AccountingSavepointCase):
+class TestAccountTax(AccountTestInvoicingCommon):
 
     @classmethod
-    def setUpClass(cls):
-        super().setUpClass()
+    def setUpClass(cls, chart_template_ref=None):
+        super().setUpClass(chart_template_ref=chart_template_ref)
 
         # By default, tests are run with the current user set on the first company.
         cls.env.user.company_id = cls.company_data['company']
